@@ -2,7 +2,7 @@ jQuery(function($) {
 
 	function runAjaxAction(button, action) {
 		button = $(button);
-		var panel = button.closest('.puc-debug-bar-panel-v4');
+		var panel = button.closest('.puc-debug-bar-panel-v5');
 		var responseBox = button.closest('td').find('.puc-ajax-response');
 
 		responseBox.text('Processing...').show();
@@ -14,19 +14,21 @@ jQuery(function($) {
 				_wpnonce: panel.data('nonce')
 			},
 			function(data) {
+				//The response contains HTML that should already be escaped in server-side code.
+				//phpcs:ignore WordPressVIPMinimum.JS.HTMLExecutingFunctions.html
 				responseBox.html(data);
 			},
 			'html'
 		);
 	}
 
-	$('.puc-debug-bar-panel-v4 input[name="puc-check-now-button"]').on('click', function() {
-		runAjaxAction(this, 'puc_v4_debug_check_now');
+	$('.puc-debug-bar-panel-v5 input[name="puc-check-now-button"]').on('click', function() {
+		runAjaxAction(this, 'puc_v5_debug_check_now');
 		return false;
 	});
 
-	$('.puc-debug-bar-panel-v4 input[name="puc-request-info-button"]').on('click', function() {
-		runAjaxAction(this, 'puc_v4_debug_request_info');
+	$('.puc-debug-bar-panel-v5 input[name="puc-request-info-button"]').on('click', function() {
+		runAjaxAction(this, 'puc_v5_debug_request_info');
 		return false;
 	});
 
@@ -34,7 +36,7 @@ jQuery(function($) {
 	// Debug Bar uses the panel class name as part of its link and container IDs. This means we can
 	// end up with multiple identical IDs if more than one plugin uses the update checker library.
 	// Fix it by replacing the class name with the plugin slug.
-	var panels = $('#debug-menu-targets').find('.puc-debug-bar-panel-v4');
+	var panels = $('#debug-menu-targets').find('.puc-debug-bar-panel-v5');
 	panels.each(function() {
 		var panel = $(this);
 		var uid = panel.data('uid');
