@@ -80,9 +80,15 @@
 			var data = [];
 
 			$( this ).closest( 'form' ).find( '[data-config-field]' ).each( function() {
+				var val = $( this ).val();
+
+				if ( $( this ).is( '[type=checkbox]' ) ) {
+					val = $( this ).is( ':checked' ) ? 1 : 0;
+				}
+
 				data.push( {
 					'name': $( this ).attr( 'name' ).replace( /^wpcf7-/, '' ).replace( /-/g, '_' ),
-					'value': $( this ).val()
+					'value': val
 				} );
 			} );
 
