@@ -14,7 +14,7 @@ wpcf7_include_module_file( 'constant-contact/doi.php' );
 add_action(
 	'wpcf7_init',
 	'wpcf7_constant_contact_register_service',
-	20, 0
+	120, 0
 );
 
 /**
