@@ -3,11 +3,10 @@
     // Exit if accessed directly.
     defined( 'ABSPATH' ) || exit;
 
-    $footer = new \Mediabit\Templates\Sections\Footer();
-    
-    echo $footer->render();
-    wp_footer(); 
 
+    wp_footer(); 
+    
 ?>
+
 </body>
 </html>
